Add ScoreAndComboSysterm

+++ Assets/Scripts/CanSliceObject.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public interface ICanSliceObject 
 {
     FruitType FruitType { get; set; }
-    float Score { get; set; }
+    int Score { get; set; }
 
     //实现Score请使用
     //[SerializeField]
